breaking: Move application logic to the server
Splits the Java application into a client and a server by moving the
model package to a microservice with APIs. EuroScope plugin will send
data directly to the microservice and the frontend client will be thin
with display logic only. This simplifies the architecture and opens for
implementing new frontends (like web based SPAs) in the future.